src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/outer-scope-lower-specificity-wins-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Inner scope rules wins, with higher specificity (v0 bug compatibility).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
